requirements.txt
prompt-toolkit==3.0.51
PyYAML==6.0.2